# Delay input type cast until necessary