--- a/include/lbann/layers/io/input/generic_input_layer.hpp
+++ b/include/lbann/layers/io/input/generic_input_layer.hpp
@@ -1135,8 +1135,11 @@ class generic_input_layer : public io_layer {
     // be the same except for overlapping width. Device copy should be
     // done with cudaMemcpy3D.
     prof_region_begin("copy-to-device", prof_colors[1], false);
-    assert0(dc::tensor::Copy(m_activations_t, input_tensor, dc::get_stream()));
+    assert0(dc::tensor::Copy(m_input_dev, input_tensor, dc::get_stream()));
     prof_region_end("copy-to-device", false);
+    prof_region_begin("cast-from-int16", prof_colors[1], false);
+    dc::tensor::Cast(m_activations_t, m_input_dev, dc::get_stream());
+    prof_region_end("copy-from-int16", false);
     // Note: no copy out for activation is necessary as the original
     // LBANN tensor is valid.
   }