--- a/src/data_readers/data_reader_cosmoflow.cpp
+++ b/src/data_readers/data_reader_cosmoflow.cpp
@@ -150,10 +150,14 @@ bool cosmoflow_reader::fetch_datum(Mat& X, int data_id, int mb_idx) {
   const short *data = data_npy.data<short>() + (size_t) offset * m_num_features;
   DataType *dest = X_v.Buffer();
 
+#ifdef LBANN_DISTCONV_COSMOFLOW_KEEP_INT16
+  std::memcpy(dest, data, sizeof(short) * m_num_features);
+#else
   // OPTIMIZE
   LBANN_OMP_PARALLEL_FOR
       for(int j = 0; j < m_num_features; j++)
         dest[j] = data[j] * m_scaling_factor_int16;
+#endif // LBANN_DISTCONV_COSMOFLOW_KEEP_INT16
   prof_region_end("fetch_datum", false);
   return true;
 }