… scripts/build_lbann_lc.sh