This is a place holder for I/O optimization by delaying casting up from int16 to float. It currently assumes the input data are in float, but by setting `InputType` as `short` would make the input data treated as `short`. 

To actually delay the type cast, let's make a temporary extension for Cosmoflow for now. I believe we should:

- Change the Cosmoflow data reader so that it does not cast from int16 to float but just passes the numpy data to Hydrogen matrices. At this point, input Hydrogen matrices would have correct size properties, but the latter half of its data would be just garbage.
- Change `InputType` to `short`.

@oyamay Can you do the above remaining extension?