@naoyam 2019-04-16
I implemented a temporary optimization just for Cosmoflow. This *only* works with the Cosmoflow data reader. It's opt-in, so nothing should be impacted unless explicitly enabled with the CMake option.

I haven't yet confirmed its correctness, but the performance improvement looks good now. For scattering one sample of 512^3 dataset to 8 GPUs:

- Packing: 51 ms -> 26 ms
- Scattering: 96 ms -> 34 ms
- Unpacking: 6 ms -> 3 ms

I'll merge this once the resulting MSE is confirmed to be fine.