@naoyam 2019-04-17
Looks like learning fine.